Revert "Merge branch 'add-twitter-1.1-plus' into feature-graph-info-print"
This reverts commit 0c6b312d89b44e92ed7ad5eb1ac0aeed72dff92b, reversing
changes made to aabd5d7d428a775ae9e0b0570cfd5bd19d362afa.

+++ Stoker/StokerBlower.m
@@ -17,7 +17,7 @@
 - (NSString *) description
 {
 	return [NSString stringWithFormat: @"StokerBlower: name = %@, id = %@, state = %d, on count = %ld, plot count = %ld", 
-			self.deviceName, self.deviceID, self.state, (long)self.onCount, [self.plotData count]];
+			self.deviceName, self.deviceID, self.state, self.onCount, [self.plotData count]];
 }
 
 @end
